Audit: the usage screen's "total" was not a total
The usage screen loaded the most recent 2,000 usage records and added them
up. Past the 2,001st request that number silently stopped being a total and
began understating what the player had spent on their own key — and the read
grew with the campaign besides.

Usage is now rolled up per model and request type in SQL, and the cost
estimate is computed from those few rows. The result is exact at any length
and costs a handful of rows to read. A conformance test pins the rollup
against the record-by-record estimate it replaces.

Also removed estimatedCostUsd from UsageTotals: nothing ever wrote the
column, so the SQL SUM over it was always zero and the field was a number
that could only be wrong. ModelProfile.estimateCost now takes Long token
counts, since an aggregate can exceed what an Int holds.

+++ core/src/main/kotlin/com/unbound/core/ai/Models.kt
@@ -40,11 +40,11 @@ data class ModelProfile(
     val supportsImages: Boolean get() = ModelCapability.IMAGE_GENERATION in capabilities
     val supportsTemperature: Boolean get() = ModelCapability.TEMPERATURE in capabilities
 
-    fun estimateCost(inputTokens: Int, outputTokens: Int, cachedTokens: Int = 0): Double? {
+    fun estimateCost(inputTokens: Long, outputTokens: Long, cachedTokens: Long = 0): Double? {
         val inCost = inputCostPerMillion ?: return null
         val outCost = outputCostPerMillion ?: return null
         val cachedCost = cachedInputCostPerMillion ?: inCost
-        val fresh = (inputTokens - cachedTokens).coerceAtLeast(0)
+        val fresh = (inputTokens - cachedTokens).coerceAtLeast(0L)
         return (fresh * inCost + cachedTokens * cachedCost + outputTokens * outCost) / 1_000_000.0
     }
 }
